requirements.txt
# Dependencias principales
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.7.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
mcp[cli]==1.9.1
httpx>=0.26.0
tenacity>=8.2.0
redis>=5.0.0
sqlalchemy>=2.0.0
structlog>=24.1.0
openai==1.12.0

# Dependencias de desarrollo
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
aiohttp>=3.9.3
locust>=2.24.0 